Lib/uu.py
#! /usr/bin/env python3

# Copyright 1994 by Lance Ellinghouse
# Cathedral City, California Republic, United States of America.
#                        All Rights Reserved
# Permission to use, copy, modify, oraz distribute this software oraz its
# documentation dla any purpose oraz without fee jest hereby granted,
# provided that the above copyright notice appear w all copies oraz that
# both that copyright notice oraz this permission notice appear w
# supporting documentation, oraz that the name of Lance Ellinghouse
# nie be used w advertising albo publicity pertaining to distribution
# of the software without specific, written prior permission.
# LANCE ELLINGHOUSE DISCLAIMS ALL WARRANTIES WITH REGARD TO
# THIS SOFTWARE, INCLUDING ALL IMPLIED WARRANTIES OF MERCHANTABILITY AND
# FITNESS, IN NO EVENT SHALL LANCE ELLINGHOUSE CENTRUM BE LIABLE
# FOR ANY SPECIAL, INDIRECT OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
# WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN
# ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT
# OF OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
#
# Modified by Jack Jansen, CWI, July 1995:
# - Use binascii module to do the actual line-by-line conversion
#   between ascii oraz binary. This results w a 1000-fold speedup. The C
#   version jest still 5 times faster, though.
# - Arguments more compliant przy python standard

"""Implementation of the UUencode oraz UUdecode functions.

encode(in_file, out_file [,name, mode])
decode(in_file [, out_file, mode])
"""

zaimportuj binascii
zaimportuj os
zaimportuj sys

__all__ = ["Error", "encode", "decode"]

klasa Error(Exception):
    dalej

def encode(in_file, out_file, name=Nic, mode=Nic):
    """Uuencode file"""
    #
    # If in_file jest a pathname open it oraz change defaults
    #
    opened_files = []
    spróbuj:
        jeżeli in_file == '-':
            in_file = sys.stdin.buffer
        albo_inaczej isinstance(in_file, str):
            jeżeli name jest Nic:
                name = os.path.basename(in_file)
            jeżeli mode jest Nic:
                spróbuj:
                    mode = os.stat(in_file).st_mode
                wyjąwszy AttributeError:
                    dalej
            in_file = open(in_file, 'rb')
            opened_files.append(in_file)
        #
        # Open out_file jeżeli it jest a pathname
        #
        jeżeli out_file == '-':
            out_file = sys.stdout.buffer
        albo_inaczej isinstance(out_file, str):
            out_file = open(out_file, 'wb')
            opened_files.append(out_file)
        #
        # Set defaults dla name oraz mode
        #
        jeżeli name jest Nic:
            name = '-'
        jeżeli mode jest Nic:
            mode = 0o666
        #
        # Write the data
        #
        out_file.write(('begin %o %s\n' % ((mode & 0o777), name)).encode("ascii"))
        data = in_file.read(45)
        dopóki len(data) > 0:
            out_file.write(binascii.b2a_uu(data))
            data = in_file.read(45)
        out_file.write(b' \nend\n')
    w_końcu:
        dla f w opened_files:
            f.close()


def decode(in_file, out_file=Nic, mode=Nic, quiet=Nieprawda):
    """Decode uuencoded file"""
    #
    # Open the input file, jeżeli needed.
    #
    opened_files = []
    jeżeli in_file == '-':
        in_file = sys.stdin.buffer
    albo_inaczej isinstance(in_file, str):
        in_file = open(in_file, 'rb')
        opened_files.append(in_file)

    spróbuj:
        #
        # Read until a begin jest encountered albo we've exhausted the file
        #
        dopóki Prawda:
            hdr = in_file.readline()
            jeżeli nie hdr:
                podnieś Error('No valid begin line found w input file')
            jeżeli nie hdr.startswith(b'begin'):
                kontynuuj
            hdrfields = hdr.split(b' ', 2)
            jeżeli len(hdrfields) == 3 oraz hdrfields[0] == b'begin':
                spróbuj:
                    int(hdrfields[1], 8)
                    przerwij
                wyjąwszy ValueError:
                    dalej
        jeżeli out_file jest Nic:
            # If the filename isn't ASCII, what's up przy that?!?
            out_file = hdrfields[2].rstrip(b' \t\r\n\f').decode("ascii")
            jeżeli os.path.exists(out_file):
                podnieś Error('Cannot overwrite existing file: %s' % out_file)
        jeżeli mode jest Nic:
            mode = int(hdrfields[1], 8)
        #
        # Open the output file
        #
        jeżeli out_file == '-':
            out_file = sys.stdout.buffer
        albo_inaczej isinstance(out_file, str):
            fp = open(out_file, 'wb')
            spróbuj:
                os.path.chmod(out_file, mode)
            wyjąwszy AttributeError:
                dalej
            out_file = fp
            opened_files.append(out_file)
        #
        # Main decoding loop
        #
        s = in_file.readline()
        dopóki s oraz s.strip(b' \t\r\n\f') != b'end':
            spróbuj:
                data = binascii.a2b_uu(s)
            wyjąwszy binascii.Error jako v:
                # Workaround dla broken uuencoders by /Fredrik Lundh
                nbytes = (((s[0]-32) & 63) * 4 + 5) // 3
                data = binascii.a2b_uu(s[:nbytes])
                jeżeli nie quiet:
                    sys.stderr.write("Warning: %s\n" % v)
            out_file.write(data)
            s = in_file.readline()
        jeżeli nie s:
            podnieś Error('Truncated input file')
    w_końcu:
        dla f w opened_files:
            f.close()

def test():
    """uuencode/uudecode main program"""

    zaimportuj optparse
    parser = optparse.OptionParser(usage='usage: %prog [-d] [-t] [input [output]]')
    parser.add_option('-d', '--decode', dest='decode', help='Decode (instead of encode)?', default=Nieprawda, action='store_true')
    parser.add_option('-t', '--text', dest='text', help='data jest text, encoded format unix-compatible text?', default=Nieprawda, action='store_true')

    (options, args) = parser.parse_args()
    jeżeli len(args) > 2:
        parser.error('incorrect number of arguments')
        sys.exit(1)

    # Use the binary streams underlying stdin/stdout
    input = sys.stdin.buffer
    output = sys.stdout.buffer
    jeżeli len(args) > 0:
        input = args[0]
    jeżeli len(args) > 1:
        output = args[1]

    jeżeli options.decode:
        jeżeli options.text:
            jeżeli isinstance(output, str):
                output = open(output, 'wb')
            inaczej:
                print(sys.argv[0], ': cannot do -t to stdout')
                sys.exit(1)
        decode(input, output)
    inaczej:
        jeżeli options.text:
            jeżeli isinstance(input, str):
                input = open(input, 'rb')
            inaczej:
                print(sys.argv[0], ': cannot do -t z stdin')
                sys.exit(1)
        encode(input, output)

jeżeli __name__ == '__main__':
    test()
